Kan oprette database
